huge refactor for managing game state. lots of things broken now

+++ lib/src/game_widgets/guesses_info_widget.dart
@@ -3,7 +3,7 @@ import 'package:lexpedition/src/game_data/constants.dart';
 
 class GuessesInformationWidget extends StatefulWidget {
   final int currentGuesses;
-  final double averageGuesses;
+  final int averageGuesses;
   final int bestAttempt;
 
   const GuessesInformationWidget(
